Changed (All Games): cvar "msg-uptime" specifies time in seconds (was tics) and can now be set via the HUD options menu.

+++ doomsday/plugins/jdoom/include/d_config.h
@@ -174,7 +174,7 @@ typedef struct jdoom_config_s {
 
     int             msgCount;
     float           msgScale;
-    int             msgUptime;
+    float           msgUptime;
     int             msgBlink;
     int             msgAlign;
     byte            msgShow;
