+++ doomsday/plugins/jheretic/include/h_config.h
@@ -168,7 +168,7 @@ typedef struct jheretic_config_s {
 
     int             msgCount;
     float           msgScale;
-    int             msgUptime;
+    float           msgUptime;
     int             msgBlink;
     int             msgAlign;
     byte            msgShow;
